Split the modifiers page templates into a folder with mirrored unit tests
The 414-line modifiers.tsx carried its whole surface under the mutation
gate as one blob. Split it by concept into src/ui/templates/admin/modifiers/,
each file paired with a mirror unit test under test/ui/templates/admin/modifiers/
so mutation runs per small file (seconds, not minutes) and stays focused:

- links.tsx    — the scope/answer link editors + SCOPE_LINK_KINDS
- values.ts    — ruleSummary + the form-field value maps (pure)
- aggregates.tsx — the running-totals section + recalculate page
- pages.tsx    — the list / create / edit / delete pages

Each split file is now at 100% mutation coverage against its own mirror test.
Adds a shared testModifier() fixture factory so the mirror tests don't each
re-spell the Modifier literal. The single caller now imports from the split
files; the old monolith and its orphan test are removed.
